move common utils
